+++ release-notes.md
@@ -2,12 +2,15 @@
 
 # Release notes
 ## Upcoming release
-* Updated TMXTile to 1.5.9 to fix duplicate tile index properties.
+* On MacOS, assets are now unpacked into the game folder instead of the resources folder for
+  consistency with other platforms.
+* Fixed duplicate tile index properties in some cases.
+* Fixed unpack error on MacOS with Steam.
 
 ## 1.0.1
 Released 21 November 2020.
 
-* Updated TMXTile to 1.5.8 to fix `.tmx` map files losing tile index properties.
+* Fixed `.tmx` map files losing tile index properties.
 
 ## 1.0
 Released 04 October 2020.